# openapi.yaml
# =====================================================================
#  STARTER SKELETON — Meeting 2, Lab L2
#
#  This file lints clean as it stands. That is its only purpose: to
#  prove your toolchain works before you have anything to lose.
#
#  Every line marked TODO is yours to replace with YOUR TEAM'S DOMAIN.
#  Do not submit a file that still says "Thing". A specification that
#  still contains this skeleton's nouns has not been done.
#
#  Work through 02-lab-L2-at-home.md alongside this file.
# =====================================================================

openapi: 3.1.0

info:
  # TODO: your domain.
  title: Badminton Court Booking API
  # The version of this DOCUMENT. Not the same as the v1 in the URL.
  version: 0.3.0
  description: |
    This server allows students and other users to view and book badminton courts.
    All money amounts are whole numbers in the smallest unit of the currency.
    ALl timestamps are RFC 3339 with a timezone offset.
    CLients MUST ignore response fields they do not recognise, and MUST not crash when they see a status value they do not know.
    TODO: one sentence — what this service is for, and for whom.

    ## Conventions

    All monetary amounts are integers in the minor unit of the stated
    currency; the currency is carried explicitly beside every amount.
    All timestamps are RFC 3339 with an explicit UTC offset, for
    example 2026-03-04T12:10:04+07:00.

    Clients MUST ignore unknown response fields. Clients MUST treat an
    unrecognised enumeration value as the nearest non-terminal state
    and MUST NOT crash on it. These two sentences are what make every
    additive change to this interface legal.

    ## Idempotency

    The Idempotency-Key header is required for booking requests, and its value must be a version-4 UUID with hyphens.
    Itâ€™s required on POST /v1/bookings, because creating a badminton court booking is a consequential operation that must not happen twice.
    Keys are kept for 24 hours. A key reused after that period is treated as new.
    The server returns 409 Conflict with the problem type https://api.example.com/problems/idempotency-key-reuse.

    ## State machine

    TODO: the transition table from Worksheet W4. Any pair of statuses
    not listed there is illegal and returns 409 illegal-transition.

servers:
  - url: https://api.example.com/v1
    description: Production — does not exist yet. That is Meeting 3.
  - url: http://localhost:8080/v1
    description: Local development

tags:
  - name: Things
    description: This API manages badminton court bookings and provides operations for creating, retrieving, and managing bookings.


paths:

  # -------------------------------------------------------------------
  # A collection GET with a filter parameter. Required for L2.
  # Rule 2: removing a query parameter must give you MORE results of the
  # same kind, never a DIFFERENT kind of thing.
  # -------------------------------------------------------------------
  /courts:
    get:
      operationId: listCourts
      summary: List badminton courts
      tags: [Courts]
      description: |
        TODO: the obligations live here, not in the summary. Say what
        is included and what is not — for example, whether closed or
        inactive items appear in the list, so that clients can display
        them greyed out rather than silently omitting them.
      parameters:
        - name: status
          in: query
          required: false
          description: Filter to things in one status.
          schema:
            type: string
            enum: [active, retired]
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          required: false
          description: Opaque cursor from a previous page's nextCursor.
          schema: { type: string }
      responses:
        '200':
          description: A page of badminton courts.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Court' }
                  nextCursor:
                    type: string
                    description: Absent on the last page.
                    example: eyJvIjoyMH0
        '400': { $ref: '#/components/responses/BadRequest' }
  /bookings:
    # -----------------------------------------------------------------
    # The unsafe, consequential operation. Required for L2.
    # The idempotency key is declared REQUIRED, so it is impossible to
    # read this document and not know about it.
    # -----------------------------------------------------------------
    post:
      operationId: createBooking
      summary: Create a badminton court booking
      tags: [Bookings]
      description: |
        Creates a badminton court booking.

        This is a consequential operation: the same booking intent MUST NOT
        create multiple bookings when the client retries the request.
        Clients MUST generate an Idempotency-Key when the user confirms
        the booking and MUST reuse the same key for retries of that intent.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
          example: 0f7c1b9e-3d21-4a6f-9c05-8e2b7d41a9f0
          description: |
            Generated by the client at the moment the user confirms the
            action, and reused unchanged for every retry of that same
            intent. Retained 24 hours. Reuse with a different body
            returns 409 idempotency-key-reuse.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewBooking' }
      responses:
        '201':
          description: |
            Created — or, on a retry with the same key and the same
            body, the replayed original response.
          headers:
            Location:
              description: URI of the created booking.
              schema: { type: string }
              # Without this example the mock server serves the literal
              # word "string" in the header. Every example you omit is a
              # place where the mock stops being useful to a client
              # developer — which is the whole point of having one.
              example: /v1/bookings/bkg_51Fa93cD
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }

  /courts/{courtId}:
    parameters:
      - name: courtId
        in: path
        required: true
        description: Opaque, server-generated. Never a database integer.
        schema:
          type: string
          pattern: '^crt_[A-Za-z0-9]{3,}$'
        example: crt_51Fa93cD
    get:
      operationId: getCourt
      summary: Fetch one badminton court
      tags: [Courts]
      responses:
        '200':
          description: The badminton court.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Court' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }

  # -------------------------------------------------------------------
  # Rule 5 — a state transition is a sub-resource, and its URI is a
  # NOUN. Not /markRetired, not PATCH {"status":"retired"}.
  # -------------------------------------------------------------------
  /courts/{courtId}/retirement:
    parameters:
      - name: courtId
        in: path
        required: true
        schema:
          type: string
          pattern: '^crt_[A-Za-z0-9]{3,}$'
        example: crt_51Fa93cD
    post:
      operationId: deactivateCourt
      summary: Retire a badminton court
      tags: [Courts]
      description: |
        TODO: your own transition. Note that the transition carries a
        body of its own (the reason), a response of its own (who and
        when), and can be read back later with GET — none of which a
        verb in the URI would give you.

        Re-sending this when the thing is ALREADY retired returns 200
        with the existing record, not 409. The requested end state
        already holds and the caller's intent has been satisfied.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  example: Replaced by a newer revision
      responses:
        '200':
          description: Already retired. The existing record is returned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Retirement' }
        '201':
          description: Retired.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Retirement' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: |
            The transition is not permitted from the current status.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example:
                type: https://api.example.com/problems/illegal-transition
                title: That status change is not permitted
                status: 409
                detail: crt_51Fa93cD is active; retirement is refused.
                instance: /v1/courts/crt_51Fa93cD/retirement
                from: active
                to: retired
                allowedFrom: [active]
components:

  schemas:

    # ---------------------------------------------------------------
    # TODO: replace with your resources from Worksheet W1. You need at
    # least THREE, each with a required array and an example on EVERY
    # property. A blank example is what the Meeting 7 partner team
    # will complain about.
    # ---------------------------------------------------------------
    Court:
      type: object
      required: [id, name, location, courtType, isAvailable, status]
      properties:
        id:
          type: string
          pattern: '^crt_[A-Za-z0-9]{3,}$'
          description: Opaque, server-generated, stable forever.
          example: crt_51Fa93cD
        name:
          type: string
          example: Court 1
        location:
          type: string
          example: UGM Sports Hall
        courtType:
          type: string
          enum: [indoor, outdoor]
          example: indoor
        isAvailable:
          type: boolean
          example: true
        status:
          type: string
          enum: [active, retired]
          description: Lifecycle status of the badminton court.
          example: active

    NewBooking:
      type: object
      required: [courtId, startTime, endTime]
      properties:
        courtId:
          type: string
          pattern: '^crt_[A-Za-z0-9]{3,}$'
          description: ID of the badminton court to book.
          example: crt_51Fa93cD
        startTime:
          type: string
          format: date-time
          description: Start time of the badminton court booking.
          example: '2026-08-29T19:00:00+07:00'
        endTime:
          type: string
          format: date-time
          description: End time of the badminton court booking.
          example: '2026-08-29T20:00:00+07:00'

    Booking:
      type: object
      required: [id, courtId, startTime, endTime, status]
      properties:
        id:
          type: string
          pattern: '^bkg_[A-Za-z0-9]{3,}$'
          description: Opaque, server-generated, stable forever.
          example: bkg_51Fa93cD
        courtId:
          type: string
          description: ID of the booked badminton court.
          example: crt_51Fa93cD
        startTime:
          type: string
          format: date-time
          description: Start time of the booking.
          example: '2026-08-29T19:00:00+07:00'
        endTime:
          type: string
          format: date-time
          description: End time of the booking.
          example: '2026-08-29T20:00:00+07:00'
        status:
          type: string
          enum: [confirmed, cancelled]
          example: confirmed

    Retirement:
      type: object
      required: [courtId, reason, retiredAt]
      properties:
        courtId:
          type: string
          pattern: '^crt_[A-Za-z0-9]{3,}$'
          example: crt_51Fa93cD
        reason:
          type: string
          example: Replaced by a newer revision
        retiredAt:
          type: string
          format: date-time
          example: '2026-03-04T12:14:00+07:00'

    # ---------------------------------------------------------------
    # RFC 9457 Problem Details. Defined ONCE, referenced everywhere,
    # so that a client able to parse one of your errors can parse all
    # of them. Extension members go alongside these five.
    # ---------------------------------------------------------------
    Problem:
      type: object
      required: [type, title, status]
      # RFC 9457 permits EXTENSION MEMBERS: your own fields alongside the
      # five below. They are what turns a merely correct error into an
      # interface a client can build a good experience on, and they are
      # explicitly assessed in A1 — so the schema must allow them.
      # Remove this line and your own extension members become lint
      # errors, which is exactly how a team ends up with useless errors.
      additionalProperties: true
      properties:
        type:
          type: string
          format: uri
          description: |
            Stable URI identifying the KIND of problem. Client code
            branches on this exact string, so it must never change
            once published.
          example: https://api.example.com/problems/illegal-transition
        title:
          type: string
          description: |
            Short, human-readable, FIXED for a given type. Written for
            a developer reading a log, not for an end user.
          example: That status change is not permitted
        status:
          type: integer
          description: The HTTP status, repeated so the document stays
            meaningful when logged away from its response.
          example: 409
        detail:
          type: string
          description: |
            Specific to this occurrence. Never a stack trace, an SQL
            fragment, or an internal hostname.
          example: thg_51Fa93cD is draft; retirement is refused.
        instance:
          type: string
          description: |
            The URI of this specific occurrence. In Chapter 9 the
            correlation identifier goes here.
          example: /v1/things/thg_51Fa93cD/retirement

  responses:

    BadRequest:
      description: The request was malformed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.example.com/problems/malformed-request
            title: The request could not be parsed
            status: 400

    NotFound:
      description: No such resource — or none you may see.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.example.com/problems/not-found
            title: Resource not found
            status: 404

    Conflict:
      description: |
        The request was valid but refused by a business rule. This is a
        domain OUTCOME, not a server error. Do not retry it.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.example.com/problems/court-slot-unavailable
            title: The badminton court is already booked for this time slot
            status: 409

    UnprocessableEntity:
      description: Well-formed but semantically invalid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.example.com/problems/validation-failed
            title: One or more fields are invalid
            status: 422