# service/src/routes/courts.js
const express = require("express");
const { findCourtById, findAllCourts } = require("../store/courts");
const { toCourtRepresentation } = require("../representations/courts");
const { problem } = require("../problem");

const router = express.Router();

router.get("/courts", async (req, res) => {
  const allowedQueryParameters = new Set(["status", "limit", "cursor"]);
  const unknownQueryParameter = Object.keys(req.query).find(
    (parameter) => !allowedQueryParameters.has(parameter)
  );

  if (unknownQueryParameter) {
    return problem(res, 400, "malformed-request", {
      detail: `Unknown query parameter: ${unknownQueryParameter}`
    });
  }

  const { status, limit, cursor } = req.query;
  let decodedCursor;

  // An empty cursor is schema-valid and means the first page.
  if (cursor !== undefined && cursor !== "") {
    const decoded = Buffer.from(cursor, "base64").toString("utf8");
    decodedCursor = /^crt_[A-Za-z0-9]{3,}$/.test(decoded) ? decoded : undefined;
  }

  let parsedLimit = 20;

  if (limit !== undefined) {
    parsedLimit = Number(limit);

    if (
      !Number.isInteger(parsedLimit) ||
      parsedLimit < 1 ||
      parsedLimit > 100
    ) {
      return problem(res, 400, "malformed-request", { detail: "Invalid limit value" });
    }
  }

  // Checked for presence, not truthiness: "" is not in the documented enum.
  if (status !== undefined && !["active", "retired"].includes(status)) {
    return problem(res, 400, "malformed-request", { detail: "Invalid status value" });
  }

  // Work
  const courts = await findAllCourts(
    status,
    parsedLimit,
    decodedCursor
  );

  const hasNextPage = courts.length > parsedLimit;

  if (hasNextPage) {
    courts.pop();
  }



  // Representation
  const items = courts.map(toCourtRepresentation);

  const nextCursor =
    hasNextPage
      ? Buffer.from(courts[courts.length - 1].id).toString("base64")
      : undefined;

  return res.status(200).json({
    items,
    ...(nextCursor && { nextCursor })
  });
});

router.get("/courts/:courtId", async (req, res) => {
  const { courtId } = req.params;

  // Validation
  const courtIdPattern = /^crt_[A-Za-z0-9]{3,}$/;

  if (!courtIdPattern.test(courtId)) {
    return problem(res, 400, "malformed-request", { detail: "Invalid courtId format" });
  }

  // Work
  const court = await findCourtById(courtId);

  if (!court) {
    return problem(res, 404, "not-found", { detail: "Court not found" });
  }

  // Representation
  return res.status(200).json(toCourtRepresentation(court));
});

module.exports = router;